migrate only useful fn from `load_data.py `into` evaluate.py`
